// overlay/src/client/message_schema.ts
export class ColorDto {
    public r: number;
    public g: number;
    public b: number;
    constructor(params: {
        r: number,
        g: number,
        b: number,
    }) {
        this.r = params.r;
        this.g = params.g;
        this.b = params.b;
    }
    public toJsonValue(): JsonObject{
        return {
            r: this.r,
            g: this.g,
            b: this.b,
        }
    }
    public static fromJsonValue(value: JsonValue): ColorDto | Error{
        return parse_as_ColorDto(value)    }
}
// ...
export class UrlDto {
    public datascheme: 'precomputed' | undefined;
    public protocol: 'http' | 'https' | 'file' | 'memory';
    public hostname: string;
    public port: number | undefined;
    public path: string;
    public search: { [key: string]: string } | undefined;
    public fragment: string | undefined;
    constructor(params: {
        datascheme: 'precomputed' | undefined,
        protocol: 'http' | 'https' | 'file' | 'memory',
        hostname: string,
        port: number | undefined,
        path: string,
        search: { [key: string]: string } | undefined,
        fragment: string | undefined,
    }) {
        this.datascheme = params.datascheme;
        this.protocol = params.protocol;
        this.hostname = params.hostname;
        this.port = params.port;
        this.path = params.path;
        this.search = params.search;
        this.fragment = params.fragment;
    }
    public toJsonValue(): JsonObject{
        return {
            datascheme: toJsonValue(this.datascheme),
            protocol: this.protocol,
            hostname: this.hostname,
            port: toJsonValue(this.port),
            path: this.path,
            search: toJsonValue(this.search),
            fragment: toJsonValue(this.fragment),
        }
    }
    public static fromJsonValue(value: JsonValue): UrlDto | Error{
        return parse_as_UrlDto(value)    }
}
// ...
export class DataSourceDto {
    public url: UrlDto;
    public interval: Interval5DDto;
    public tile_shape: Shape5DDto;
    public spatial_resolution: [number,number,number];
    constructor(params: {
        url: UrlDto,
        interval: Interval5DDto,
        tile_shape: Shape5DDto,
        spatial_resolution: [number,number,number],
    }) {
        this.url = params.url;
        this.interval = params.interval;
        this.tile_shape = params.tile_shape;
        this.spatial_resolution = params.spatial_resolution;
    }
    public toJsonValue(): JsonObject{
        return {
            url: this.url.toJsonValue(),
            interval: this.interval.toJsonValue(),
            tile_shape: this.tile_shape.toJsonValue(),
            spatial_resolution: [this.spatial_resolution[0],
this.spatial_resolution[1],
this.spatial_resolution[2]],
        }
    }
    public static fromJsonValue(value: JsonValue): DataSourceDto | Error{
        return parse_as_DataSourceDto(value)    }
}
// ...
export class RawDataViewDto {
    public name: string;
    public url: UrlDto;
    public datasources: Array<DataSourceDto>;
    constructor(params: {
        name: string,
        url: UrlDto,
        datasources: Array<DataSourceDto>,
    }) {
        this.name = params.name;
        this.url = params.url;
        this.datasources = params.datasources;
    }
    public toJsonValue(): JsonObject{
        return {
            name: this.name,
            url: this.url.toJsonValue(),
            datasources: this.datasources.map(item => item.toJsonValue()),
        }
    }
    public static fromJsonValue(value: JsonValue): RawDataViewDto | Error{
        return parse_as_RawDataViewDto(value)    }
}
// ...
export class StrippedPrecomputedViewDto {
    public name: string;
    public url: UrlDto;
    public datasource: DataSourceDto;
    constructor(params: {
        name: string,
        url: UrlDto,
        datasource: DataSourceDto,
    }) {
        this.name = params.name;
        this.url = params.url;
        this.datasource = params.datasource;
    }
    public toJsonValue(): JsonObject{
        return {
            name: this.name,
            url: this.url.toJsonValue(),
            datasource: this.datasource.toJsonValue(),
        }
    }
    public static fromJsonValue(value: JsonValue): StrippedPrecomputedViewDto | Error{
        return parse_as_StrippedPrecomputedViewDto(value)    }
}
// ...
export class PredictionsViewDto {
    public name: string;
    public url: UrlDto;
    public raw_data: DataSourceDto;
    public classifier_generation: number;
    constructor(params: {
        name: string,
        url: UrlDto,
        raw_data: DataSourceDto,
        classifier_generation: number,
    }) {
        this.name = params.name;
        this.url = params.url;
        this.raw_data = params.raw_data;
        this.classifier_generation = params.classifier_generation;
    }
    public toJsonValue(): JsonObject{
        return {
            name: this.name,
            url: this.url.toJsonValue(),
            raw_data: this.raw_data.toJsonValue(),
            classifier_generation: this.classifier_generation,
        }
    }
    public static fromJsonValue(value: JsonValue): PredictionsViewDto | Error{
        return parse_as_PredictionsViewDto(value)    }
}
// ...
export class FailedViewDto {
    public name: string;
    public url: UrlDto;
    public error_message: string;
    constructor(params: {
        name: string,
        url: UrlDto,
        error_message: string,
    }) {
        this.name = params.name;
        this.url = params.url;
        this.error_message = params.error_message;
    }
    public toJsonValue(): JsonObject{
        return {
            name: this.name,
            url: this.url.toJsonValue(),
            error_message: this.error_message,
        }
    }
    public static fromJsonValue(value: JsonValue): FailedViewDto | Error{
        return parse_as_FailedViewDto(value)    }
}
// ...
export class UnsupportedDatasetViewDto {
    public name: string;
    public url: UrlDto;
    constructor(params: {
        name: string,
        url: UrlDto,
    }) {
        this.name = params.name;
        this.url = params.url;
    }
    public toJsonValue(): JsonObject{
        return {
            name: this.name,
            url: this.url.toJsonValue(),
        }
    }
    public static fromJsonValue(value: JsonValue): UnsupportedDatasetViewDto | Error{
        return parse_as_UnsupportedDatasetViewDto(value)    }
}
// ...
export class ViewerAppletStateDto {
    public frontend_timestamp: number;
    public data_views: Array<RawDataViewDto | StrippedPrecomputedViewDto | FailedViewDto | UnsupportedDatasetViewDto>;
    public prediction_views: Array<PredictionsViewDto>;
    public label_colors: Array<ColorDto>;
    constructor(params: {
        frontend_timestamp: number,
        data_views: Array<RawDataViewDto | StrippedPrecomputedViewDto | FailedViewDto | UnsupportedDatasetViewDto>,
        prediction_views: Array<PredictionsViewDto>,
        label_colors: Array<ColorDto>,
    }) {
        this.frontend_timestamp = params.frontend_timestamp;
        this.data_views = params.data_views;
        this.prediction_views = params.prediction_views;
        this.label_colors = params.label_colors;
    }
    public toJsonValue(): JsonObject{
        return {
            frontend_timestamp: this.frontend_timestamp,
            data_views: this.data_views.map(item => toJsonValue(item)),
            prediction_views: this.prediction_views.map(item => item.toJsonValue()),
            label_colors: this.label_colors.map(item => item.toJsonValue()),
        }
    }
    public static fromJsonValue(value: JsonValue): ViewerAppletStateDto | Error{
        return parse_as_ViewerAppletStateDto(value)    }
}
// ...
export class MakeDataViewParams {
    public view_name: string;
    public url: UrlDto;
    constructor(params: {
        view_name: string,
        url: UrlDto,
    }) {
        this.view_name = params.view_name;
        this.url = params.url;
    }
    public toJsonValue(): JsonObject{
        return {
            view_name: this.view_name,
            url: this.url.toJsonValue(),
        }
    }
    public static fromJsonValue(value: JsonValue): MakeDataViewParams | Error{
        return parse_as_MakeDataViewParams(value)    }
}